+++ source/operators/rekognition/check_rekognition_status.py
