+++ source/operators/rekognition/start_rekognition.py
